* Disabled setting of uninterested when no more chunks can be requested, this needs to be fixed.
* Minor adjustments to the usage of rak::priority_queue_insert.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/rtorrent@611 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/display/utils.cc
@@ -76,7 +76,7 @@ print_ddhhmm(char* buf, unsigned int length, time_t t) {
   unsigned int s;
 
   if (t / (24 * 3600) < 100)
-    s = snprintf(buf, length, "%2li:%02li:%02li", t / (24 * 3600), (t / 3600) % 24, (t / 60) % 60);
+    s = snprintf(buf, length, "%2i:%02i:%02i", (int)t / (24 * 3600), ((int)t / 3600) % 24, ((int)t / 60) % 60);
   else
     s = snprintf(buf, length, "--:--:--");
